# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/tenants": {
      "post": {
        "operationId": "create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTenantDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedTenantResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "platform": []
          }
        ],
        "summary": "Create a tenant and its owner API key",
        "tags": [
          "Platform"
        ]
      }
    },
    "/v1/tenants/{id}/api-keys": {
      "post": {
        "operationId": "issueKey",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueApiKeyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssuedApiKeyResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "platform": []
          }
        ],
        "summary": "Issue an additional API key for a tenant",
        "tags": [
          "Platform"
        ]
      }
    },
    "/v1/tenant": {
      "get": {
        "operationId": "me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "The tenant the API key belongs to",
        "tags": [
          "Tenant"
        ]
      },
      "patch": {
        "operationId": "update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTenantDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Update settings, including the dunning policy",
        "tags": [
          "Tenant"
        ]
      }
    },
    "/v1/tenant/api-keys": {
      "get": {
        "operationId": "listKeys",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApiKeyResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "List API keys (prefixes only)",
        "tags": [
          "Tenant"
        ]
      }
    },
    "/v1/tenant/api-keys/{id}": {
      "delete": {
        "operationId": "revoke",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Revoke an API key",
        "tags": [
          "Tenant"
        ]
      }
    },
...
    "/v1/invoices": {
      "post": {
        "operationId": "create",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Any unique string; a repeat with the same key returns the first response instead of acting twice",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvoiceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Create a draft invoice",
        "tags": [
          "Invoices"
        ]
      },
      "get": {
        "operationId": "list",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Opaque cursor from a previous page",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "draft",
                "issued",
                "partially_paid",
                "paid",
                "void"
              ]
            }
          },
          {
            "name": "customerId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "overdue",
            "required": false,
            "in": "query",
            "description": "Only open invoices past their due date",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoicePage"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "List invoices, newest first",
        "tags": [
          "Invoices"
        ]
      }
    },
    "/v1/invoices/{id}": {
      "get": {
        "operationId": "get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Get an invoice with its lines and dunning state",
        "tags": [
          "Invoices"
        ]
      }
    },
    "/v1/invoices/{id}/issue": {
      "post": {
        "operationId": "issue",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Any unique string; a repeat with the same key returns the first response instead of acting twice",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueInvoiceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Issue a draft: assign the next number, set the due date, send it",
        "tags": [
          "Invoices"
        ]
      }
    },
    "/v1/invoices/{id}/void": {
      "post": {
        "operationId": "void",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoidInvoiceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Void an unpaid invoice",
        "tags": [
          "Invoices"
        ]
      }
    },
    "/v1/invoices/{id}/pdf": {
      "get": {
        "operationId": "pdf",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "The invoice as a PDF; rendered on first request if the worker has not yet",
        "tags": [
          "Invoices"
        ]
      }
    },
    "/v1/invoices/{invoiceId}/payments": {
      "post": {
        "operationId": "record",
        "parameters": [
          {
            "name": "invoiceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordPaymentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Record a payment against an invoice",
        "tags": [
          "Payments"
        ]
      },
      "get": {
        "operationId": "list",
        "parameters": [
          {
            "name": "invoiceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PaymentResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Payments recorded against an invoice, oldest first",
        "tags": [
          "Payments"
        ]
      }
    },
...
    "/v1/invoices/{id}/dunning": {
      "get": {
        "operationId": "notices",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DunningNoticeResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Notices sent for an invoice",
        "tags": [
          "Dunning"
        ]
      }
    },
...
  },
  "info": {
    "title": "Dunlin",
    "description": "Multi-tenant invoicing, dunning and payment reconciliation. Authenticate with a tenant API key as a Bearer token; tenant creation uses the platform token.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "api-key": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Tenant API key"
      },
      "platform": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Platform-Token"
      }
    },
    "schemas": {
      "CreateTenantDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Acme Treuhand AG"
          },
          "country": {
            "type": "string",
            "example": "CH",
            "description": "ISO 3166-1 alpha-2"
          },
          "currency": {
            "type": "string",
            "example": "CHF",
            "description": "ISO 4217; every invoice of the tenant uses it"
          },
          "invoicePrefix": {
            "type": "string",
            "example": "INV",
            "default": "INV"
          },
          "paymentTermsDays": {
            "type": "number",
            "example": 30,
            "default": 14,
            "minimum": 0,
            "maximum": 365
          },
          "dunningPolicy": {
            "type": "object",
            "description": "Escalation schedule; see the Dunning section. Defaults to reminder at 3 days, second reminder with a 15.00 fee at 14, final notice with a 40.00 fee at 30, 5 % annual interest.",
            "additionalProperties": true
          }
        },
        "required": [
          "name",
          "country",
          "currency"
        ]
      },
      "TenantResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "invoicePrefix": {
            "type": "string"
          },
          "paymentTermsDays": {
            "type": "number"
          },
          "dunningPolicy": {
            "type": "object",
            "additionalProperties": true
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "country",
          "currency",
          "invoicePrefix",
          "paymentTermsDays",
          "dunningPolicy",
          "createdAt"
        ]
      },
      "IssuedApiKeyResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "owner",
              "accountant",
              "viewer"
            ]
          },
          "keyPrefix": {
            "type": "string",
            "description": "The first characters of the key, for identification"
          },
          "createdAt": {
            "type": "string"
          },
          "lastUsedAt": {
            "type": "object",
            "nullable": true
          },
          "revokedAt": {
            "type": "object",
            "nullable": true
          },
          "key": {
            "type": "string",
            "description": "The full key. Shown once; it is not stored."
          }
        },
        "required": [
          "id",
          "name",
          "role",
          "keyPrefix",
          "createdAt",
          "lastUsedAt",
          "revokedAt",
          "key"
        ]
      },
      "CreatedTenantResponse": {
        "type": "object",
        "properties": {
          "tenant": {
            "$ref": "#/components/schemas/TenantResponse"
          },
          "apiKey": {
            "$ref": "#/components/schemas/IssuedApiKeyResponse"
          }
        },
        "required": [
          "tenant",
          "apiKey"
        ]
      },
      "IssueApiKeyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Accounting integration"
          },
          "role": {
            "type": "string",
            "enum": [
              "owner",
              "accountant",
              "viewer"
            ],
            "example": "accountant"
          }
        },
        "required": [
          "name",
          "role"
        ]
      },
      "UpdateTenantDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Acme Treuhand AG"
          },
          "invoicePrefix": {
            "type": "string",
            "example": "INV"
          },
          "paymentTermsDays": {
            "type": "number",
            "example": 30,
            "minimum": 0,
            "maximum": 365
          },
          "dunningPolicy": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "ApiKeyResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "owner",
              "accountant",
              "viewer"
            ]
          },
          "keyPrefix": {
            "type": "string",
            "description": "The first characters of the key, for identification"
          },
          "createdAt": {
            "type": "string"
          },
          "lastUsedAt": {
            "type": "object",
            "nullable": true
          },
          "revokedAt": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "role",
          "keyPrefix",
          "createdAt",
          "lastUsedAt",
          "revokedAt"
        ]
      },
...
      "CreateInvoiceDto": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "string",
            "format": "uuid"
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceLineDto"
            }
          },
          "notes": {
            "type": "string",
            "example": "Payable within 30 days."
          }
        },
        "required": [
          "customerId",
          "lines"
        ]
      },
...
      "InvoiceResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "number": {
            "type": "object",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "issued",
              "partially_paid",
              "paid",
              "void"
            ]
          },
          "customerId": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "issuedOn": {
            "type": "object",
            "nullable": true
          },
          "dueOn": {
            "type": "object",
            "nullable": true
          },
          "subtotalMinor": {
            "type": "number"
          },
          "taxMinor": {
            "type": "number"
          },
          "totalMinor": {
            "type": "number"
          },
          "paidMinor": {
            "type": "number"
          },
          "balanceMinor": {
            "type": "number"
          },
          "overdue": {
            "type": "boolean",
            "description": "Open and past due, as of today"
          },
          "notes": {
            "type": "object",
            "nullable": true
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceLineResponse"
            }
          },
          "dunning": {
            "$ref": "#/components/schemas/DunningSummaryResponse"
          },
          "voidedAt": {
            "type": "object",
            "nullable": true
          },
          "voidReason": {
            "type": "object",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "number",
          "status",
          "customerId",
          "currency",
          "issuedOn",
          "dueOn",
          "subtotalMinor",
          "taxMinor",
          "totalMinor",
          "paidMinor",
          "balanceMinor",
          "overdue",
          "notes",
          "lines",
          "dunning",
          "voidedAt",
          "voidReason",
          "createdAt",
          "updatedAt"
        ]
      },
      "InvoicePage": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceResponse"
            }
          },
          "nextCursor": {
            "type": "object",
            "nullable": true,
            "description": "Pass as ?cursor= for the next page; null on the last page"
          }
        },
        "required": [
          "items",
          "nextCursor"
        ]
      },
      "IssueInvoiceDto": {
        "type": "object",
        "properties": {
          "issuedOn": {
            "type": "string",
            "example": "2026-09-01",
            "description": "Defaults to today; the due date follows from the payment terms"
          }
        }
      },
      "VoidInvoiceDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "example": "Issued to the wrong customer"
          }
        }
      },
      "RecordPaymentDto": {
        "type": "object",
        "properties": {
          "amountMinor": {
            "type": "number",
            "example": 12500,
            "description": "Minor units; at most the open balance"
          },
          "receivedOn": {
            "type": "string",
            "example": "2026-09-15"
          },
          "method": {
            "type": "string",
            "enum": [
              "bank_transfer",
              "card",
              "cash",
              "other"
            ]
          },
          "reference": {
            "type": "string",
            "example": "Bank ref 8812"
          }
        },
        "required": [
          "amountMinor",
          "receivedOn",
          "method"
        ]
      },
      "PaymentResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "invoiceId": {
            "type": "string"
          },
          "amountMinor": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "receivedOn": {
            "type": "string"
          },
          "method": {
            "type": "string",
            "enum": [
              "bank_transfer",
              "card",
              "cash",
              "other"
            ]
          },
          "reference": {
            "type": "object",
            "nullable": true
          },
          "source": {
            "type": "string",
            "enum": [
              "manual",
              "bank_import"
            ]
          },
          "bankImportLineId": {
            "type": "object",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          },
          "invoiceStatus": {
            "type": "string",
            "description": "The invoice after this payment"
          },
          "invoiceBalanceMinor": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "invoiceId",
          "amountMinor",
          "currency",
          "receivedOn",
          "method",
          "reference",
          "source",
          "bankImportLineId",
          "createdAt",
          "invoiceStatus",
          "invoiceBalanceMinor"
        ]
      },
...
    }
  }
}